- remove map class - update ros2idl to convert to messagedefinitions - lowercase IDL - separate internal types
