+++ packages/omgidl-parser/src/IdlNodes/IdlNode.ts
@@ -1,21 +1,15 @@
 import { ConstantValue } from "@foxglove/message-definition";
 
-import { IDLDefinitionMap, toScopedIdentifier } from "./IDLDefinitionMap";
-import { SIMPLE_TYPES, normalizeType } from "./primitiveTypes";
-import {
-  BaseASTNode,
-  ConstantASTNode,
-  EnumASTNode,
-  IDLMessageDefinition,
-  IDLMessageDefinitionField,
-} from "./types";
+import { BaseASTNode, ConstantASTNode, EnumASTNode } from "../astTypes";
+import { SIMPLE_TYPES, normalizeType } from "../primitiveTypes";
+import { IDLMessageDefinition, IDLMessageDefinitionField } from "../types";
 
-export class IDLNode<T extends BaseASTNode = BaseASTNode> implements BaseASTNode {
-  private map: IDLDefinitionMap;
+export class IdlNode<T extends BaseASTNode = BaseASTNode> implements BaseASTNode {
+  private map: Map<string, IdlNode>;
   protected readonly astNode: T;
   readonly scopePath: string[];
 
-  constructor(scopePath: string[], astNode: T, idlMap: IDLDefinitionMap) {
+  constructor(scopePath: string[], astNode: T, idlMap: Map<string, IdlNode>) {
     this.scopePath = scopePath;
     this.astNode = astNode;
     this.map = idlMap;
@@ -41,8 +35,12 @@ export class IDLNode<T extends BaseASTNode = BaseASTNode> implements BaseASTNode
     return toScopedIdentifier([...this.scopePath, this.name]);
   }
 
-  protected getNode(scopePath: string[], name: string): IDLNode {
-    const maybeNode = this.map.getNode(scopePath, name);
+  protected getNode(scopePath: string[], name: string): IdlNode {
+    const maybeNode = resolveScopedOrLocalNodeReference({
+      usedIdentifier: name,
+      scopeOfUsage: scopePath,
+      definitionMap: this.map,
+    });
     if (maybeNode == undefined) {
       throw new Error(
         `Could not find node ${name} in ${scopePath.join("::")} referenced by ${
@@ -53,18 +51,18 @@ export class IDLNode<T extends BaseASTNode = BaseASTNode> implements BaseASTNode
     return maybeNode;
   }
 
-  protected getConstantNode(identifier: string): IDLConstantNode {
+  protected getConstantNode(identifier: string): ConstantIdlNode {
     const maybeConstantNode = this.getNode(this.scopePath, identifier);
-    if (!(maybeConstantNode instanceof IDLConstantNode)) {
+    if (!(maybeConstantNode instanceof ConstantIdlNode)) {
       throw new Error(`Expected ${this.name} to be a constant in ${this.scopedIdentifier}`);
     }
     return maybeConstantNode;
   }
 }
 
-export class IDLConstantNode extends IDLNode<ConstantASTNode> {
+export class ConstantIdlNode extends IdlNode<ConstantASTNode> {
   private needsResolution = false;
-  constructor(scopePath: string[], astNode: ConstantASTNode, idlMap: IDLDefinitionMap) {
+  constructor(scopePath: string[], astNode: ConstantASTNode, idlMap: Map<string, IdlNode>) {
     super(scopePath, astNode, idlMap);
     if (!SIMPLE_TYPES.has(astNode.type)) {
       this.needsResolution = true;
@@ -78,11 +76,11 @@ export class IDLConstantNode extends IDLNode<ConstantASTNode> {
     return this.astNode.type;
   }
 
-  private referencedEnumNode?: IDLEnumNode = undefined;
-  private getReferencedEnumNode(): IDLEnumNode {
+  private referencedEnumNode?: EnumIdlNode = undefined;
+  private getReferencedEnumNode(): EnumIdlNode {
     if (this.referencedEnumNode == undefined) {
       const maybeEnumNode = this.getNode(this.scopePath, this.astNode.type);
-      if (!(maybeEnumNode instanceof IDLEnumNode)) {
+      if (!(maybeEnumNode instanceof EnumIdlNode)) {
         throw new Error(`Expected ${this.astNode.type} to be an enum in ${this.scopedIdentifier}`);
       }
       this.referencedEnumNode = maybeEnumNode;
@@ -113,8 +111,8 @@ export class IDLConstantNode extends IDLNode<ConstantASTNode> {
   }
 }
 
-export class IDLEnumNode extends IDLNode<EnumASTNode> {
-  constructor(scopePath: string[], astNode: EnumASTNode, idlMap: IDLDefinitionMap) {
+export class EnumIdlNode extends IdlNode<EnumASTNode> {
+  constructor(scopePath: string[], astNode: EnumASTNode, idlMap: Map<string, IdlNode>) {
     super(scopePath, astNode, idlMap);
   }
 
@@ -141,7 +139,7 @@ export class IDLEnumNode extends IDLNode<EnumASTNode> {
     return undefined;
   }
 
-  get enumerators(): IDLConstantNode[] {
+  get enumerators(): ConstantIdlNode[] {
     return this.astNode.enumerators.map((enumerator) =>
       this.getConstantNode(toScopedIdentifier([...this.scopePath, this.name, enumerator])),
     );
@@ -158,4 +156,36 @@ export class IDLEnumNode extends IDLNode<EnumASTNode> {
       aggregatedKind: "module",
     };
   }
+}
+
+function resolveScopedOrLocalNodeReference({
+  usedIdentifier,
+  scopeOfUsage,
+  definitionMap,
+}: {
+  usedIdentifier: string;
+  scopeOfUsage: string[];
+  definitionMap: Map<string, IdlNode>;
+}): IdlNode | undefined {
+  // If using local un-scoped identifier, it will not be found in the definitions map
+  // In this case we try by building up the namespace prefix until we find a match
+  let referencedNode = undefined;
+  const namespacePrefixes = [...scopeOfUsage];
+  const currPrefix: string[] = [];
+  for (;;) {
+    const identifierToTry = toScopedIdentifier([...currPrefix, usedIdentifier]);
+    referencedNode = definitionMap.get(identifierToTry);
+    if (referencedNode != undefined) {
+      break;
+    }
+    if (namespacePrefixes.length === 0) {
+      break;
+    }
+    currPrefix.push(namespacePrefixes.shift()!);
+  }
+
+  return referencedNode;
+}
+function toScopedIdentifier(path: string[]): string {
+  return path.join("::");
 }
