+++ packages/omgidl-parser/src/IdlNodes/index.ts
@@ -1,0 +1,4 @@
+export * from "./IdlNode";
+export * from "./ModuleIdlNode";
+export * from "./ReferenceTypeIdlNode";
+export * from "./StructIdlNode";
