+++ packages/omgidl-parser/src/parseIdl.ts
@@ -1,8 +1,5 @@
-import { MessageDefinition } from "@foxglove/message-definition";
-
-import { IDLDefinitionMap } from "./IDLDefinitionMap";
-import { IDLNodeProcessor } from "./IDLNodeProcessor";
 import { parseIdlToAST } from "./parseIdlToAST";
+import { buildMap, toIDLMessageDefinitions } from "./processIdl";
 import { IDLMessageDefinition } from "./types";
 
 /**
@@ -13,23 +10,6 @@ import { IDLMessageDefinition } from "./types";
 export function parseIdl(messageDefinition: string): IDLMessageDefinition[] {
   const rawIdlDefinitions = parseIdlToAST(messageDefinition);
 
-  const idlProcessor = new IDLDefinitionMap(rawIdlDefinitions);
-  return idlProcessor.toIDLMessageDefinitions();
-}
-
-/**
- * Parses IDL schema to flattened MessageDefinitions that can be used to serialize/deserialize messages
- * @param messageDefinition - idl decoded message definition string
- * @returns - parsed message definition
- */
-export function parseIdlToMessageDefinition(messageDefinition: string): MessageDefinition[] {
-  const rawIdlDefinitions = parseIdlToAST(messageDefinition);
-
-  const idlProcessor = new IDLNodeProcessor(rawIdlDefinitions);
-  idlProcessor.resolveEnumTypes();
-  idlProcessor.resolveConstantUsage();
-  idlProcessor.resolveTypeDefComplexity();
-  idlProcessor.resolveStructMember();
-
-  return idlProcessor.toMessageDefinitions();
+  const idlMap = buildMap(rawIdlDefinitions);
+  return toIDLMessageDefinitions(idlMap);
 }
